Moved interface to the models folder
