+++ src/api/backend.api.ts
@@ -1,0 +1,1 @@
+export const getChallengeFromApi = () => {};
