+++ src/models/cliOptions.models.ts
@@ -1,0 +1,11 @@
+export interface ICliOptions {
+  title?: String;
+  edabitId?: String;
+  author?: String;
+  authorId?: String;
+  tags?: String | String[];
+  minDifficulty?: String | Number;
+  minQuality?: String | Number;
+  programmingLanguage?: String;
+  skipConfirmation?: Boolean;
+}
