+++ src/models/index.ts
@@ -1,2 +1,3 @@
-export * from "./answerChoice";
-export * from "./consoleMessage";
+export * from "./answerChoice.models";
+export * from "./cliOptions.models";
+export * from "./consoleMessage.models";
