+++ src/utils/command.utils.ts
@@ -1,16 +1,5 @@
 import command from "commander";
-
-export interface ICliOptions {
-  title?: String;
-  edabitId?: String;
-  author?: String;
-  authorId?: String;
-  tags?: String | String[];
-  minDifficulty?: String | Number;
-  minQuality?: String | Number;
-  programmingLanguage?: String;
-  skipConfirmation?: Boolean;
-}
+import { ICliOptions } from "../models";
 
 /**
  * Parses the commoand line arguments and returns the options
